# fix(sqlite): `SqliteEventCacheStore` is 35 times faster