This patch adds a benchmark for the `LinkedChunk` with the following matrix:

* Handling 10, 100, 1_000, 10_000 and 100_000 events,
* With no store, with the `MemoryStore`, with the `SqliteEventCacheStore`.

Before this patch, the case with no store and `MemoryStore` were not a blocker. I was surprisingly delighted to see that the `LinkedChunk` is able to handle a throughput of 9.3 millions events per second. With the `MemoryStore`, it drops at 4.5 millions events per second, not bad for a test-tailored store. However, for `SqliteEventCacheStore`, the story was clearly different… **7_300 events per second**, only. And it got worst with more events.

![Screenshot 2025-02-28 at 13 42 48](https://github.com/user-attachments/assets/f666e453-141e-4035-8517-26dcc1609f24)

I dug into this for 2 days, and found several things.

In the future we can improve the benchmark to test more operations, but I wanted to solve the performance issue first.

## Prepared statements for `Update::PushNewItems`

This is addressed as a standalone patch. I've replaced a `query()` in a loop by a `prepare()` outside the loop, then an `execute()` inside the loop.

Guess what? It barely improved the situation. I've reached **8_400 events per second**. Not really the improvement I was expecting. However, the performance was now linear with the number of events. Before that, it was totally unpredictable. That's a first step.

I've tried to use a bulk insertion to avoid to call `execute()` in a loop. A single `prepare()` + a single `execute()` like so:

```sql
INSERT INTO events VALUES (?, ?, ?, ?, ?), (?, ?, ?, ?, ?), …
```

Absolutely zero difference. Well. Time not well-spent on this one, but it was worth the try.

## Remove all tables: time for a new schema

The tables were fine but none of them contained primary keys. I hoped it could solve things. Sadly, [`ALTER TABLE`](https://sqlite.org/lang_altertable.html) doesn't allow to change the schema of tables, only a bit of renamings, and column changes.

The solution is to remove all indexes and all tables, and to recreate them. Same table names, same column names, but different semantics:

* `linked_chunks` has a primary key over `room_id` and `id` (the chunk identifier), this pair is necessarily unique (this uniqueness constraint was missing before),
* `gaps` has a primary key over `room_id` and `chunk_id`, same reason,
* `events` has a primary key composed of `room_id`, `chunk_id` and `position`. Again, this tuple is necessarily unique, and this constraint was missing. The uniqueness on `event_id` has been added in https://github.com/matrix-org/matrix-rust-sdk/pull/4685.

Why `events` cannot use `event_id` as its primary key? Because an `event_id` can be null if the event is invalid. After all, it's allowed in SQLite (because it was a bug in the past, buggy behaviours are kept).

> [!NOTE]
> Why an invalid event is stored in the database? It's still a mystery to me.
> cc @bnjbvr 

With this new schema, things have improved a bit. **14_000 events per second**, better!

## `WITHOUT ROWID`

Please read [Clustered Indexes and the `WITHOUT ROWID` Optimization](https://www.sqlite.org/withoutrowid.html).

In addition to the new schema, the 3 tables are marked as `WITHOUT ROWID`. It has dramatically improved the performance. We are reaching **178_000 events per second**!

I lied a bit in the previous section. At first, the primary key for `events` was `event_id`, and it did improved the performance, but `WITHOUT ROWID` expects the primary key to _NOT_ be null. So I had to find another unique key, hence the tuple `room_id`, `chunk_id` and `position`. And this whole re-designed was with `WITHOUT ROWID` in mind, but it's the result of several tries and failures.

## Restore the index over `events`

Finally, in the new schema, I re-create the index for `events.room_id` and `events.event_id`. And we are finally reaching **260_000 events per second**!!

## Results for 100_000 events

_(These results are a bit different because, at the time of publishing these patches, I re-run them on my laptop running on a low battery. The orders are the same though.)_

### No store

<img width="960" alt="Screenshot 2025-02-28 at 18 34 05" src="https://github.com/user-attachments/assets/74cea1e6-a6ea-4101-b058-eb3ffc66e34b" />

|            | Lower bound    | Estimate       | Upper bound    |
|------------|----------------|----------------|----------------|
| Slope      | 13.641 ms      | 15.264 ms      | 16.948 ms      |
| Throughput | 5.9003 Melem/s | 6.5513 Melem/s | 7.3308 Melem/s |
| R²         | 0.6328193      | 0.7358459      | 0.6261084      |
| Mean       | 13.285 ms      | 14.181 ms      | 15.422 ms      |
| Std. Dev.  | 548.23 µs      | 1.8684 ms      | 2.7520 ms      |
| Median     | 13.077 ms      | 13.206 ms      | 14.839 ms      |
| MAD        | 42.535 µs      | 299.56 µs      | 2.2939 ms      |

### `MemoryStore`

<img width="965" alt="Screenshot 2025-02-28 at 18 34 14" src="https://github.com/user-attachments/assets/f109e23e-4022-4f68-8e06-8f9bab61374b" />

|            | Lower bound    | Estimate       | Upper bound    |
|------------|----------------|----------------|----------------|
| Slope      | 32.587 ms      | 33.022 ms      | 33.414 ms      |
| Throughput | 2.9928 Melem/s | 3.0283 Melem/s | 3.0687 Melem/s |
| R²         | 0.9932222      | 0.9961436      | 0.9937653      |
| Mean       | 32.545 ms      | 32.748 ms      | 33.022 ms      |
| Std. Dev.  | 102.11 µs      | 417.19 µs      | 600.38 µs      |
| Median     | 32.490 ms      | 32.601 ms      | 32.891 ms      |
| MAD        | 50.689 µs      | 166.82 µs      | 458.06 µs      |

### `SqliteEventCacheStore`

<img width="973" alt="Screenshot 2025-02-28 at 18 34 21" src="https://github.com/user-attachments/assets/e858fa11-55a2-459d-803a-5b8baec44ebd" />

|            | Lower bound    | Estimate       | Upper bound    |
|------------|----------------|----------------|----------------|
| Slope      | 403.39 ms      | 403.88 ms      | 404.57 ms      |
| Throughput | 247.17 Kelem/s | 247.60 Kelem/s | 247.90 Kelem/s |
| R²         | 0.9998947      | 0.9999231      | 0.9998668      |
| Mean       | 403.77 ms      | 404.39 ms      | 405.02 ms      |
| Std. Dev.  | 643.36 µs      | 1.0718 ms      | 1.2887 ms      |
| Median     | 403.47 ms      | 404.15 ms      | 405.54 ms      |
| MAD        | 205.61 µs      | 1.3011 ms      | 1.8636 ms      |

## Cons

Migrating from version 5 to 6 will erase all existing caches. I consider it's okay since (i) it's a cache, (ii) it's not released yet.

---

* Address https://github.com/matrix-org/matrix-rust-sdk/issues/3280